Use unix timestamps instead of DateTime
To represent times and dates an unix timestamp is better as it protects
against DateTime object mutation, and is what Syfmony Sessions uses in
the MetadataBag.

The lastRequest is renamed to lastUsed to align it with the MetadataBag
terminology, and now, the lastUsed timestamp can be set from outside to
reuse the MetadataBag info.